Move cgroup tests to cgroup_test.go
