+++ pkg/security/tests/container_test.go
@@ -13,7 +13,6 @@ import (
 	"os"
 	"os/exec"
 	"strconv"
-	"syscall"
 	"testing"
 	"time"
 
@@ -225,54 +224,4 @@ func createCGroup(name string) (string, error) {
 	}
 
 	return cgroupPath, nil
-}
-func TestCGroupID(t *testing.T) {
-	if testEnvironment == DockerEnvironment {
-		t.Skip("skipping cgroup ID test in docker")
-	}
-
-	SkipIfNotAvailable(t)
-
-	ruleDefs := []*rules.RuleDefinition{
-		{
-			ID:         "test_container_flags",
-			Expression: `open.file.path == "{{.Root}}/test-open" && cgroup.id == "cg1"`,
-		},
-	}
-	test, err := newTestModule(t, nil, ruleDefs)
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer test.Close()
-
-	cgroupPath, err := createCGroup("cg1")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.RemoveAll(cgroupPath)
-
-	testFile, testFilePtr, err := test.Path("test-open")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	test.Run(t, "cgroup-id", func(t *testing.T, kind wrapperType, cmdFunc func(cmd string, args []string, envs []string) *exec.Cmd) {
-		test.WaitSignal(t, func() error {
-			fd, _, errno := syscall.Syscall6(syscall.SYS_OPENAT, 0, uintptr(testFilePtr), syscall.O_CREAT, 0711, 0, 0)
-			if errno != 0 {
-				return error(errno)
-			}
-			return syscall.Close(int(fd))
-
-		}, func(event *model.Event, rule *rules.Rule) {
-			assertTriggeredRule(t, rule, "test_container_flags")
-			assertFieldEqual(t, event, "open.file.path", testFile)
-			assertFieldEqual(t, event, "container.id", "")
-			assertFieldEqual(t, event, "container.runtime", "")
-			assert.Equal(t, containerutils.CGroupFlags(0), event.CGroupContext.CGroupFlags)
-			assertFieldEqual(t, event, "cgroup.id", "cg1")
-
-			test.validateOpenSchema(t, event)
-		})
-	})
 }
